src/main/java/com/live/code/spring_boot_mockito/CustomerController.java
package com.live.code.spring_boot_mockito;

import lombok.AllArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

@RestController
@AllArgsConstructor
public class CustomerController {

    private final CustomerService customerService;

    @GetMapping("/customer")
    public ResponseEntity<Customer> getCustomerById(@RequestParam String id) {
        return ResponseEntity.ok(customerService.getCustomer(id));
    }
}
